+++ src/pages/_app.tsx
@@ -2,12 +2,15 @@ import "@/styles/globals.css";
 
 import { ThemeProvider } from "@emotion/react";
 import type { AppProps } from "next/app";
+import { useState } from "react";
 
-import theme from "@/styles/theme";
+import { darkTheme, lightTheme } from "@/styles/theme";
 
 export default function App({ Component, pageProps }: AppProps) {
+  const [isDark, setIsDark] = useState(false);
+
   return (
-    <ThemeProvider theme={theme}>
+    <ThemeProvider theme={isDark ? darkTheme : lightTheme}>
       <Component {...pageProps} />
     </ThemeProvider>
   );
